+++ API/src/main/java/fr/maxlego08/essentials/api/convert/Convert.java
@@ -1,0 +1,9 @@
+package fr.maxlego08.essentials.api.convert;
+
+import org.bukkit.command.CommandSender;
+
+public interface Convert {
+
+    void convert(CommandSender sender);
+
+}