+++ API/src/main/java/fr/maxlego08/essentials/api/vault/PlayerVaults.java
@@ -26,4 +26,6 @@ public interface PlayerVaults {
     Vault getVault(int vaultId);
 
     Optional<Vault> find(ItemStack currentItem);
+
+    Vault firstAvailableVault();
 }